implement performance calculation and beatmap downloading

+++ pkg/osu/types.go
@@ -153,17 +153,17 @@ type UserScores []struct {
 		Status           string      `json:"status"`
 		TotalLength      int         `json:"total_length"`
 		Version          string      `json:"version"`
-		Accuracy         float64         `json:"accuracy"`
-		Ar               float64        `json:"ar"`
+		Accuracy         float64     `json:"accuracy"`
+		Ar               float64     `json:"ar"`
 		BeatmapsetID     int         `json:"beatmapset_id"`
-		Bpm              int         `json:"bpm"`
+		Bpm              float64         `json:"bpm"`
 		Convert          bool        `json:"convert"`
 		CountCircles     int         `json:"count_circles"`
 		CountSliders     int         `json:"count_sliders"`
 		CountSpinners    int         `json:"count_spinners"`
-		Cs               float64         `json:"cs"`
+		Cs               float64     `json:"cs"`
 		DeletedAt        interface{} `json:"deleted_at"`
-		Drain            float64        `json:"drain"`
+		Drain            float64     `json:"drain"`
 		HitLength        int         `json:"hit_length"`
 		IsScoreable      bool        `json:"is_scoreable"`
 		LastUpdated      time.Time   `json:"last_updated"`
@@ -202,7 +202,7 @@ type UserScores []struct {
 		Video          bool        `json:"video"`
 	} `json:"beatmapset"`
 	Weight struct {
-		Percentage float64     `json:"percentage"`
+		Percentage float64 `json:"percentage"`
 		Pp         float64 `json:"pp"`
 	} `json:"weight"`
 	User struct {
